+++ mod/beacon/blockchain/processor/errors.go
@@ -18,7 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package blockchain
+package processor
 
 import "github.com/berachain/beacon-kit/mod/errors"
