--- mod/node-core/pkg/services/blockchain/processor.go
@@ -1,44 +1,0 @@
-// SPDX-License-Identifier: BUSL-1.1
-//
-// Copyright (C) 2024, Berachain Foundation. All rights reserved.
-// Use of this software is governed by the Business Source License included
-// in the LICENSE file of this repository and at www.mariadb.com/bsl11.
-//
-// ANY USE OF THE LICENSED WORK IN VIOLATION OF THIS LICENSE WILL AUTOMATICALLY
-// TERMINATE YOUR RIGHTS UNDER THIS LICENSE FOR THE CURRENT AND ALL OTHER
-// VERSIONS OF THE LICENSED WORK.
-//
-// THIS LICENSE DOES NOT GRANT YOU ANY RIGHT IN ANY TRADEMARK OR LOGO OF
-// LICENSOR OR ITS AFFILIATES (PROVIDED THAT YOU MAY USE A TRADEMARK OR LOGO OF
-// LICENSOR AS EXPRESSLY REQUIRED BY THIS LICENSE).
-//
-// TO THE EXTENT PERMITTED BY APPLICABLE LAW, THE LICENSED WORK IS PROVIDED ON
-// AN “AS IS” BASIS. LICENSOR HEREBY DISCLAIMS ALL WARRANTIES AND CONDITIONS,
-// EXPRESS OR IMPLIED, INCLUDING (WITHOUT LIMITATION) WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
-// TITLE.
-
-package blockchain
-
-import (
-	"context"
-
-	"github.com/berachain/beacon-kit/mod/primitives/pkg/transition"
-)
-
-// Processor is the interface for the blockchain processor.
-//
-//nolint:lll // long types
-type Processor[
-	BeaconBlockT BeaconBlock[BeaconBlockBodyT, ExecutionPayloadT],
-	BeaconBlockBodyT BeaconBlockBody[ExecutionPayloadT],
-	DepositT any,
-	ExecutionPayloadT ExecutionPayload,
-	ExecutionPayloadHeaderT ExecutionPayloadHeader,
-	GenesisT Genesis[DepositT, ExecutionPayloadHeaderT],
-] interface {
-	ProcessBeaconBlock(ctx context.Context, blk BeaconBlockT) (transition.ValidatorUpdates, error)
-	ProcessGenesisData(ctx context.Context, genesisData GenesisT) (transition.ValidatorUpdates, error)
-	ReceiveBlock(ctx context.Context, blk BeaconBlockT) error
-	VerifyIncomingBlock(ctx context.Context, blk BeaconBlockT) error
-}
