--- mod/node-core/pkg/services/validator/types.go
@@ -1,119 +1,0 @@
-// SPDX-License-Identifier: BUSL-1.1
-//
-// Copyright (C) 2024, Berachain Foundation. All rights reserved.
-// Use of this software is governed by the Business Source License included
-// in the LICENSE file of this repository and at www.mariadb.com/bsl11.
-//
-// ANY USE OF THE LICENSED WORK IN VIOLATION OF THIS LICENSE WILL AUTOMATICALLY
-// TERMINATE YOUR RIGHTS UNDER THIS LICENSE FOR THE CURRENT AND ALL OTHER
-// VERSIONS OF THE LICENSED WORK.
-//
-// THIS LICENSE DOES NOT GRANT YOU ANY RIGHT IN ANY TRADEMARK OR LOGO OF
-// LICENSOR OR ITS AFFILIATES (PROVIDED THAT YOU MAY USE A TRADEMARK OR LOGO OF
-// LICENSOR AS EXPRESSLY REQUIRED BY THIS LICENSE).
-//
-// TO THE EXTENT PERMITTED BY APPLICABLE LAW, THE LICENSED WORK IS PROVIDED ON
-// AN “AS IS” BASIS. LICENSOR HEREBY DISCLAIMS ALL WARRANTIES AND CONDITIONS,
-// EXPRESS OR IMPLIED, INCLUDING (WITHOUT LIMITATION) WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
-// TITLE.
-
-package validator
-
-import (
-	"context"
-
-	gethprimitives "github.com/berachain/beacon-kit/mod/geth-primitives"
-	"github.com/berachain/beacon-kit/mod/primitives/pkg/common"
-	"github.com/berachain/beacon-kit/mod/primitives/pkg/constraints"
-	"github.com/berachain/beacon-kit/mod/primitives/pkg/crypto"
-	"github.com/berachain/beacon-kit/mod/primitives/pkg/eip4844"
-	"github.com/berachain/beacon-kit/mod/primitives/pkg/math"
-)
-
-// BeaconBlock represents a beacon block interface.
-type BeaconBlock[
-	AttestationDataT any,
-	BeaconBlockT any,
-	BeaconBlockBodyT BeaconBlockBody[
-		AttestationDataT, DepositT, Eth1DataT, ExecutionPayloadT, SlashingInfoT,
-	],
-	DepositT,
-	Eth1DataT,
-	ExecutionPayloadT,
-	SlashingInfoT any,
-] interface {
-	constraints.SSZMarshallable
-	// NewWithVersion creates a new beacon block with the given parameters.
-	NewWithVersion(
-		slot math.Slot,
-		proposerIndex math.ValidatorIndex,
-		parentBlockRoot common.Root,
-		forkVersion uint32,
-	) (BeaconBlockT, error)
-	// GetSlot returns the slot of the beacon block.
-	GetSlot() math.Slot
-	// GetParentBlockRoot returns the parent block root of the beacon block.
-	GetParentBlockRoot() common.Root
-	// SetStateRoot sets the state root of the beacon block.
-	SetStateRoot(common.Root)
-	// GetStateRoot returns the state root of the beacon block.
-	GetStateRoot() common.Root
-	// GetBody returns the body of the beacon block.
-	GetBody() BeaconBlockBodyT
-}
-
-// BeaconBlockBody represents a beacon block body interface.
-type BeaconBlockBody[
-	AttestationDataT, DepositT, Eth1DataT, ExecutionPayloadT, SlashingInfoT any,
-] interface {
-	constraints.SSZMarshallable
-	constraints.Nillable
-	// SetRandaoReveal sets the Randao reveal of the beacon block body.
-	SetRandaoReveal(crypto.BLSSignature)
-	// SetEth1Data sets the Eth1 data of the beacon block body.
-	SetEth1Data(Eth1DataT)
-	// SetDeposits sets the deposits of the beacon block body.
-	SetDeposits([]DepositT)
-	// SetExecutionPayload sets the execution data of the beacon block body.
-	SetExecutionPayload(ExecutionPayloadT)
-	// SetGraffiti sets the graffiti of the beacon block body.
-	SetGraffiti(common.Bytes32)
-	// SetAttestations sets the attestations of the beacon block body.
-	SetAttestations([]AttestationDataT)
-	// SetSlashingInfo sets the slashing info of the beacon block body.
-	SetSlashingInfo([]SlashingInfoT)
-	// SetBlobKzgCommitments sets the blob KZG commitments of the beacon block
-	// body.
-	SetBlobKzgCommitments(eip4844.KZGCommitments[gethprimitives.ExecutionHash])
-}
-
-// Eth1Data represents the eth1 data interface.
-type Eth1Data[T any] interface {
-	// New creates a new eth1 data with the given parameters.
-	New(
-		depositRoot common.Root,
-		depositCount math.U64,
-		blockHash gethprimitives.ExecutionHash,
-	) T
-}
-
-// EventPublisher represents the event publisher interface.
-type EventPublisher[T any] interface {
-	// PublishEvent publishes an event.
-	Publish(context.Context, T) error
-}
-
-type EventFeed[T any] interface {
-	Subscribe() (chan T, error)
-}
-
-// SlotData represents the slot data interface.
-type SlotData[AttestationDataT, SlashingInfoT any] interface {
-	// GetSlot returns the slot of the incoming slot.
-	GetSlot() math.Slot
-	// GetAttestationData returns the attestation data of the incoming slot.
-	GetAttestationData() []AttestationDataT
-	// GetSlashingInfo returns the slashing info of the incoming slot.
-	GetSlashingInfo() []SlashingInfoT
-}
